requirements.txt
openai==1.7.2
feedparser==6.0.10
pre-commit==3.5.0
requests==2.31.0

#
streamlit==1.30.0
#
langchain==0.1.0
langchainhub==0.1.14
langchain_openai==0.0.2.post1
langchain-community==0.0.12
# paper downloading
scholarly==1.7.11
biopython==1.81
# requests-html==0.10.0
json-repair==0.4.5
metapub==0.5.5
PyPaperBot==1.2.2
PyPDF2==3.0.1

# db
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
faiss-cpu==1.7.4
tiktoken==0.5.2

# utils
fuzzywuzzy==0.18.0
pyrootutils==1.0.4
